Refactoring cogs
