+++ commands/voiceCommands.py
@@ -1,0 +1,6 @@
+# Description: Cog that contains commands that affect in some way the voice channels
+
+from discord.ext import commands
+
+class voiceCommands(commands.cog):
+
